requirements.txt
paddlepaddle
paddleocr
opencv-python
vietocr
matplotlib
uvicorn==0.18.2
python-multipart==0.0.5
fastapi==0.78.0
pandas